move note to its own file
