+++ lib/zoho_hub/base_record.rb
@@ -251,12 +251,5 @@ module ZohoHub
     end
   end
 
-  class Note < BaseRecord
-    attributes :id, :created_by, :modified_by, :owner, :parent_id, :created_time, :voice_note,
-               :note_title, :note_content
-
-    attribute_translation id: :id
-    alias title note_title
-    alias content note_content
-  end
+  require 'zoho_hub/modules/note'
 end
